+++ core/core-database/src/main/java/com/baka3k/core/database/model/CastEntity.kt
@@ -47,7 +47,7 @@ data class CastEntity(
     val creditId: String,
 
     @ColumnInfo(defaultValue = "0")
-    val movieId: Int,
+    val movieId: Long,
 )
 
 fun CastEntity.asExternalModel() = Cast(
