+++ core/core-network/src/main/java/com/baka3k/core/network/datasource/CreditNetworkDataSource.kt
@@ -1,9 +1,8 @@
 package com.baka3k.core.network.datasource
 
 import com.baka3k.core.common.result.Result
-import com.baka3k.core.model.PagingInfo
 import com.baka3k.core.network.model.NetworkCreditsResponse
 
 interface CreditNetworkDataSource {
-    suspend fun getCredits(movieId: Int): Result<NetworkCreditsResponse>
+    suspend fun getCredits(movieId: Long): Result<NetworkCreditsResponse>
 }
