+++ features/feature-movie-detail-compose/src/main/java/com/baka3k/test/movie/detail/interactor/GetCrewUseCase.kt
@@ -10,10 +10,10 @@ import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetCrewUseCase @Inject constructor(
-) : UpStreamSingleUseCaseParameter<Int, Flow<Result<List<Crew>>>> {
+) : UpStreamSingleUseCaseParameter<Long, Flow<Result<List<Crew>>>> {
     @Inject
     lateinit var creditRepository: CreditRepository
-    override fun invoke(idMovie: Int): Flow<Result<List<Crew>>> {
+    override fun invoke(idMovie: Long): Flow<Result<List<Crew>>> {
         return creditRepository.getCrewStream(idMovie).asResult()
     }
 }
